configgadgetwidget: defer creating subwidgets for 3s
This allows the app to open a couple seconds quicker.  The trade off is,
if you try to go to config gadget widget too fast, you encounter a
couple second delay while we catch up and instantiate all the widgets.
But not everyone will go to config gadget first, and not all will go
there extremely quickly after loading.