pios_servo/oneshot: Ensure dead time.
Now max pulse length is passed into pios_servo, and we make sure we're
"past" that point by a defined margin before we'll disable the timer.

pios_servo should get factored common in the future and non-HPWM
variants of the code should get removed in that refactor.

This depends on all channels in a oneshot bank being defined with the
same "max" time to be fully protective.