Throttle curve dilutes AH/PH altitude control
During altitude hold and position hold, throttle_desired is used to
maintain the current altitude. However, the throttle_desired value is
desensitized when the throttle curve is applied in acctuator.c. In an
attempt to overcome this desentization, one can crank up the attitude
compensation as well as the Kp/Kd terms which make uav setup different
between similar crafts with and without throttle curveis. Furthermore, when
the user exceeds their throttle curve knees, due to controlled altitude
change with the stick or under stong winds conditions, the altitude change
can be drastic and unexpected due to the higher AH settings.

This change simply disables the throttle curve during AH and PH
flightmodes. Not sure if the proposed change is the best solution, so
interested in nay feedback.