naze32: configurable interrupt pin for mpu6050
The pin moves about among the various naze incarnations and isn't as
simple to detect as "is this a rev 5?"  Afromini Amaze is effectively a
rev 4, but has the MPU on port C.

This allows the user to override that when the default isn't working
correctly.  My Amaze works now.

Closes #702