stabilization settings: turn up weak leveling rate
We suggest using weakleveling a lot to learn rate, but it needs to be
more effective to do so.  The current coefficients do almost nothing--
.1 corresponds to taking out 63% of the pitch/roll in 10 seconds, or 30
seconds to remove 95% of the bank.

New value of .33 takes out 63% of the pitch/roll in 3 seconds, or 9
seconds to get 95% of the way to level.

Put another way, at 45 degree roll, it'll want to initially level at 15
deg/sec.  With 150 deg/sec default full-stick deflection rates, it'll
take 10% deflection to hold that attitude.