gcs: uploader: simplify buttons
I need the UI space, and now the buttons always have conceptually the
same purpose regardless of state.

The leftmost button is Rescue / Enter bootloader.  It means to do what's
necessary to get us into the loader.

The next button is Boot / Reboot.  It means to do what's necessary to
get a fresh instance of the firmware booted.