flight: cc3d: lower settings partition size to 256k
Note this has the risk of losing existing settings during upgrade in the
event that there's been a lot of updates and the active arena has spilled
more than 4 times.   But this is a one-time penalty in trade for making
upgrade muuuuuuch faster than downloading a 1MB file, and I don't think
many people will have done the required 1024 object saves to spill that
far.  (figure complete config is ~16 objects on CC3D, so 64 complete
config saves).