circqueue: refactor to allow multi-elem operations
Previously circqueue only allowed one operation at a time.  Now we
allow the caller to fill in / consume many at once.

This is not sufficient for pios_com, because the concept of number
available is a bit different.  Here it's the number contiguous, but
you may very well be able to ultimately write more elements without
reader activity (or vice versa).  New calls -- or another parameter
signalling out-- will be required to meet this use case.