logviewer: fix a bug where you can't add new plots
Reported by @ufodziner --- if you have "closed" a plot area by popping
it out and closing the window, then it is often impossible to add new
ones because of how we lay out the widgets.  Fix this-- pick a new
place to put the plot when our anchor for adding new ones is not
visible.