revolution: support N25Q128 NOR flash.
None of the additional capacity is used-- this is merely to tolerate
this part.  Reported OK by yds on forum here:
https://forum.dronin.org/forum/d/87-readytoflyquads-revolution-acro/25

This part is shipping on some clone Revolution boards without full
sensor suite.  This is just a preliminary piece of work-- if changes are
made to support not-full sensor suites these targets can be supported.