transmitter_controL: update edge detector
If we are disarmed (e.g. from RX timeout or low-throttle timeout) but
an arming switch in armed position, we attempt to spot a positive switch
edge.

However, a failsafe is considered as the arming switch in the disarmed
position.

It's important that we not "remember" that failsafe state for the edge
detector.