gcs controller: adjust throttle range
Right now the throttle control is misleading because you have to push it
halfway up to do anything, since the bottom half represents half of the
range.  Arguably this is "right", but it's not very useful.  Change
things so that instead we use a throttle neutral value of 1100 which
corresponds to 10% of travel.