Added an airspeed panel.
This isn't using STD_PANEL because I just don't bother getting the
airspeed except in the rare cases someone needs it.  There are other
panels in this form and this could likely be generalized, but this at
least prevents us from grabbing data that's almost never going to be
present.