bootloader: change to use common vendor/prod id
Also bump bootloader rev to 0x87.

This is beneficial so that androidgcs can refuse to try connecting
to the bootloader.  It should simplify things going forward.

It will be quite some time before most of our users pick this up, so we
can't require this ID for quite some time.