Stabilization: Change default deriv gamma to 0.6 (#1545)
When there's step inputs to rates-- (easy to generate in attitude
mode)-- there can easily be a big overshoot in achieved rate even
with an otherwise good tune.  This in turn tends to ring a couple
of times and look a little underdamped.

Setting derivative gamma to 0.6 turns down the derivative for the
setpoint portion by 40%-- which significantly limits the chaos.

Many PID systems choose to not apply D to setpoint at all, but IMO
the setpoint D is helpful in quick response to big commands.  This
is a compromise value to maintain familiar characteristics but
avoid the worst case ringing on big moves.