attitude: raise AccKI, both during startup and at baseline
This is based on eyeballing actual behavior in plots from gooch/irc.  He
has a quad with a severe gyro bias issue.  The initial transient at
startup isn't fixing it, and the further trimming would take
approximately forever to correct it.

The arming value is chosen to be 15; a value of 6 would deliver a time
constant of approximately the startup interval (6 seconds); 15 is about
2/5 of the arming interval or we should expect about 92% of the error to
get trimmed out.

Since that's raising by 15x, raise the "normal" value by a factor of 8x
to 0.4.  By the above rationale that's a time constant of about 90
seconds-- enough that slow, very small drift is coped with, but small
attitude errors don't do much at all