GCS:Reporter: Fix trailing ^0, no more lies from GIT_DIRTY
GIT_TAG had the distance from the ref on the end, we already have a coarser version of this ("dirty") and it makes using the tag name more difficult. This will also now result in a blank string rather than "undefined" when HEAD isn't tagged.

Funnily enough GIT_DIRTY was broken and always returned false as the qmake function isEmpty takes a variable _name_, not a value, and of course the variable named "" is empty, as is "4 files changed, 8 insertions(+), 6 deletions(-)" if we're actually dirty.

Additionally, use plumbing commands rather than porcelain.