remove ground path follower
It's been unused in a very long time-- and doubtful whether it still
works.  It's also very far outside the premise of the project.

But most importantly, I don't think the concept is very useful-- blind
GPS-coordinate-based navigation is not a very good premise for doing
anything useful with a ground vehicle.