gcs: hwsettings: handle uint32 data
Namely, radio IDs.  If the value exceeded 2^31, it was clipped to 0.
This meant if you saved the hardware page it'd unbind your OpLink, etc.

Now, use a double spinbox to represent these values.  Though there is
still an unsightly thing of the units "hex" being displayed when the
value is decidedly not hexy.