pios: f4: flash_internal: erase flash from RAM
Need to have this one little routine in RAM, so that we can do
lengthy flash operations without hanging long enough for watchdog
to expire.

Still needs to have a way to test arming status at the outset. I'm
thinking probably a weak symbol or something like that (as this code can
be used in the loader, too).