flightgear: add drprotocol file
The necessary steps to use this are something like this:

1. Install file in flightgear install.  for me, this is
cp drprotocol.xml /usr/share/games/flightgear/Protocol/
2. Start flightgear-- command line like:
fgfs --generic=socket,out,30,127.0.0.1,4444,udp,drprotocol
--generic=socket,in,30,127.0.0.1,4445,udp,drprotocol

2.b. it's worth familiarizing yourself with flightgear some.
In particular, things I do on the menus are
Cessna 172P-Aircraft Options: and then turn on "Start with engine
running"/
Location-Position Aircraft in Air is useful, too, to drop the plane in
air at like 5000ft, though then I need to do Cessna 172P-Autostart manually.

3. Start GCS-- note that GCS and flightgear must be on the same screen
(though FG can be made into a tiny window) because it doesn't do the
physics model at full rate if it's not visible.
4. Get the aircraft into whatever starting condition you want the sim to
experience.
5. Start the sim: ./build/sim/sim.elf -g 4444
6. Connect GCS to localhost:9000, and then go through an arming sequence
and pick a flight mode.  You can fly the simulator around in rate mode,
etc.

What's still needed is a bit of work for the magnetometer and
gpsposition and gpsvelocity objects to be set based on what flightgear
is returning-- then it should be possible to fly entire fixed wing
scenarios with a good physics model.