pios_tcp: buffering improvement
Right now, com drivers, if they can't call a PIOS_COM rx callback
successfully (e.g. because the PIOS_COM buffer is full) spill the data
on the floor; for the most part this is correct-- better to spill old
data than new data, etc.

In the case of pios_tcp, the sender is adaptive in rate based on the
rate which data is consumed (thanks to TCP's windowing mechanism).  So
it's essential that instead we block trying to put it into the PIOS_COM
buffer, which in turn has the effect of slowing down the ultimate
sender.