refactor hangtime time limit to stabilization
It was previously in actuator, which is a bit of a nasty place for it.
In turn, it's complicated the path to 3D stab, and also does not use
I term stabilization when in hangtime.

Note existing configurations will be lost-- as the field is in a
different uavo now (StabilizationSettings instead of ActuatorSettings).