manualcontrol: support for 3D flight
Autodetermine whether to use collective channel for thrust based on
presence (not vehicle type).

Determine whether throttle is bidirectional based on relative position
of neutral signal.  If it's >35% of the way through the range, we're
trying for 3D.

For modes that want ascent/descent rates, ensure we always give them
a value of range -1..1, instead of 0..1.