ChibiOS,HAL,F446: New ChibiOS suddenly has issues with HSECLK being at 16000000 with the F446.
The RE1 sets the HSE clock to 16MHz. ChibiOS 17.x does some more rigorous checks on the mcuconf.h settings than old ChibiOS 2.6, and considers this invalid. We already shipped it like that, though. So I assume this will work anyway, so this check is neutered for our only F446 target.

Untested, because my RE1 is broken.