smoothcontrol: Drop extended mode, add linear interpolation, allow user-selected duty cycle.
Normal had a fix duty cycle of 50%, extended had one of 75%. Extended mode was dropped in favor of the user specifying a duty cycle. Default value is 50%.

Added linear interpolation mode, that also follows the duty cycle. This allows interpolating to the target value in a short time than the receiver updates. This would essentially slant the risers of the stair stepped control signal.

Also, removed the useless limit, and the throttle smoothing gets disabled when the signal is NaN.