gcs: autotune: compute vertical autotune
Use information on hover throttle (presuming this is 1G acceleration)
and roll/pitch tau (assuming this corresponds to thrust tau) in order to
calculate a conservative (but still more aggressive than stock tuning)
vertical tune.

Produces similar results to the defaults on things that use 50% hover
power with 65ms tau.