python: add new uavofile class to read gcs uav exports
Provide a githash and the uav file, and the uav file is parsed under
that githash.  Any unknown fields are set to their default, and XML is
written out.  The main() part of this is just a dumb proof of concept--
the intended functionality of this code is to be able to
programmatically import settings.