transmitter_control: improve collective check
Some of the users that have tripped over #2101 have had the channel set
to e.g. SBus/None.  Don't just look at the channel type, but instead
look at the channel number, to see if the "whole thing" is configured,
so that it's possible to set everything to SBus without unanticipated
consequences.

Also factor the channel valid check to a static function, which may be
desirable to use elsewhere in transmitter_control.